requirements.txt
django==5.1.6
psycopg2==2.9.10
faker==36.0.0
djangorestframework==3.15.2